added type cast

+++ stopwatch.c
@@ -10,7 +10,7 @@ void getTime(char** returnValue){
 
 	char* str;
 	str = malloc(20 * sizeof(char));
-	unsigned long long millis = tval.tv_sec * 1000 + tval.tv_usec / 1000;
+	unsigned long long millis = (unsigned long long) tval.tv_sec * 1000 + tval.tv_usec / 1000;
 	sprintf(str, "%lld", millis);
 	
 	strcpy(*returnValue, str);
